extract Galley root command to server. (#12073)
* Replace root command of Galley with server mode.

* Fix linter issue.
