+++ install/kubernetes/helm/istio/charts/galley/templates/deployment.yaml
@@ -45,6 +45,7 @@ spec:
           - containerPort: 9901
           command:
           - /usr/local/bin/galley
+          - server
           - --meshConfigFile=/etc/mesh-config/mesh
           - --livenessProbeInterval=1s
           - --livenessProbePath=/healthliveness
